refactor(emulator): route native emulator ViewModel through use cases, extract menu sheet, rename Delta -> Native
- New use cases GetDownloadedROMUseCase, ResolveROMFileUseCase and EmulatorSaveStatesUseCase replace direct repo/helper access in the view model
- NativeEmulatorSession takes the save-states use case instead of the raw store
- DeltaEmulatorView/VM/Session move to UI/Emulator/Native/ and are renamed to NativeEmulator*
- EmulatorMenuSheet plus loading/error overlays move into their own file
- Settings UI labels updated to 'Native (DeltaCore, etc.)'
